# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Attendly API",
        "description": "교회 출석 관리 시스템 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/admin/logs": {
            "get": {
                "tags": [
                    "관리자 로그 API"
                ],
                "summary": "시스템 로그 조회",
                "description": "조건별 시스템 로그를 조회합니다 (관리자 전용)",
                "operationId": "getLogs",
                "parameters": [
                    {
                        "name": "level",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startTime",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "endTime",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "timestamp,desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponseSystemLogResponseDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/logs/{id}": {
            "get": {
                "tags": [
                    "관리자 로그 API"
                ],
                "summary": "시스템 로그 단건 조회",
                "description": "특정 ID의 시스템 로그를 조회합니다 (관리자 전용)",
                "operationId": "getLogById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSystemLogResponseDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/logs/levels": {
            "get": {
                "tags": [
                    "관리자 로그 API"
                ],
                "summary": "로그 레벨 목록 조회",
                "description": "시스템에 등록된 모든 로그 레벨을 조회합니다 (관리자 전용)",
                "operationId": "getLogLevels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListString"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/logs/categories": {
            "get": {
                "tags": [
                    "관리자 로그 API"
                ],
                "summary": "로그 카테고리 목록 조회",
                "description": "시스템에 등록된 모든 로그 카테고리를 조회합니다 (관리자 전용)",
                "operationId": "getLogCategories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListString"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponsePageResponseSystemLogResponseDto": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResponseSystemLogResponseDto"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "PageResponseSystemLogResponseDto": {
                "required": [
                    "hasMore",
                    "items",
                    "totalCount"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SystemLogResponseDto"
                        }
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hasMore": {
                        "type": "boolean"
                    }
                }
            },
            "SystemLogResponseDto": {
                "required": [
                    "category",
                    "level",
                    "message",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "level": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "additionalInfo": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ipAddress": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userAgent": {
                        "type": "string"
                    },
                    "serverInstance": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseSystemLogResponseDto": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/SystemLogResponseDto"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ApiResponseListString": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}